Do not reset CMAKE_C_COMPILER
Probe the compiler's ability to handle the clang specific option
rather then hard-resetting the compiler... _after_ having
evaluated the compiler's linker flag support and w/o reevaluating.

If the compiler can't be used, emit a corresponding diagnostic
and have the user restart cmake with another compiler.

Also, add MSVC branch and remove enforcent of Debug mode when
trying to test a release build. Code should be testable the
way it gets deployed, which is in release mode.