Do not reset CXX
Fail if the compiler is insufficient, but do not overwrite CXX, ever. If it is
overwritten, it causes half the project to be compiled using one compiler and
the other half with another.

Set CXX environment variable or pass -DCMAKE_CXX_COMPILER=[...] to the
cmake command line invokation or set the variable in a preset instead.